requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0

# MCP (Model Context Protocol) client
mcp>=1.0.0
langchain-mcp-adapters>=0.1.0

# Streamlit for UI
streamlit>=1.28.0

# HTTP and async support
httpx>=0.25.0
aiohttp>=3.9.0

# Environment management
python-dotenv>=1.0.0

# JSON and data handling
pydantic>=2.0.0

# Logging and utilities
loguru>=0.7.0

# Server-Sent Events support for MCP
sseclient-py>=1.8.0